Fix DAS integral transform MPI bug in dev (#216)
Fix issue that the Fock builder (which has MPI calls) is called from a root only portion of the integral transform.
We now split the MPI communicator into a root only communicator for the offending part of the integral transform. This is passed to the single slater object. Afterwards, the single slater communicator is reset.